refactor: extract songs/releases router (PR-12, Phase 1)
